quicksight/app/requirements.txt
boto3==1.24.7
botocore==1.27.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
flask-lambda-python36==0.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.8.0
